api_gateway/requirements.txt
fastapi
uvicorn
httpx
pyyaml
websockets